bin/benchmarks/requirements.txt
tqdm>=4.66.4
nltk>=3.8.1
spacy>=3.7.5
torch>=2.1.2
torchmetrics>=1.3.0.post0
tiktoken>=0.7.0
langchain>=0.2.6
langchain_community>=0.2.6
langchain_openai==0.1.7
langchain_google_vertexai>=1.0.4
langchain_huggingface>=0.0.3
langchain_qdrant>=0.1.1
datasets>=2.16.1
transformers==4.40.1
pandas>=2.0.3
evaluate>=0.4.1
accelerate>=0.31.0
bitsandbytes>=0.43.1
faiss-cpu>=1.8.0.post1
keybert>=0.8.5
yake>=0.4.8
